Refactor resumable upload files location and structure
The resumable upload related files were moved from 'files' to a new 'resumables' package. A new class 'Resumables' was introduced to hold common functionalities such as reading resumable chunks and converting entities to DTOs, making the 'FilesService' class cleaner and easier to understand.
