+++ services/tsd-api-mock/src/main/java/no/elixir/tsdapimock/resumables/ResumableUpload.java
@@ -1,4 +1,4 @@
-package no.elixir.tsdapimock.files;
+package no.elixir.tsdapimock.resumables;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
