+++ services/tsd-api-mock/src/main/java/no/elixir/tsdapimock/resumables/ResumableUploadsResponseDto.java
@@ -1,4 +1,4 @@
-package no.elixir.tsdapimock.files.dto;
+package no.elixir.tsdapimock.resumables;
 
 import java.util.ArrayList;
 import lombok.Data;
